+++ src/subcommands/mod.rs
@@ -4,6 +4,7 @@ use start::*;
 use fill::*;
 use remove::*;
 use summary::*;
+use count::*;
 use generate_completions::*;
 
 mod subcommands;
@@ -12,4 +13,5 @@ mod start;
 mod fill;
 mod remove;
 mod summary;
+mod count;
 mod generate_completions;