+++ src/subcommands/subcommands.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 use clap::{Clap, crate_version, crate_description};
 use enum_dispatch::enum_dispatch;
-use super::{Fill, Log, Remove, Start, Summary, GenerateCompletions};
+use super::{Fill, Log, Remove, Start, Summary, Count, GenerateCompletions};
 
 #[enum_dispatch]
 pub trait Subcommand {
@@ -17,6 +17,7 @@ pub enum Subcommands {
     Start(Start),
     Fill(Fill),
     Remove(Remove),
+    Count(Count),
     Summary(Summary),
     GenerateCompletions(GenerateCompletions)
 }